+++ projects/js-packages/storybook/storybook/projects.js
@@ -7,6 +7,7 @@ const projects = [
 	'../../components/components',
 	'../../connection/components',
 	'../../idc/components',
+	'../../social-logos/src/react',
 	'../../shared-extension-utils/src',
 	'../../../packages/my-jetpack/_inc/components',
 	'../../../packages/search/src/dashboard/components',