+++ projects/packages/connection/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.11.0] - 2024-06-26
+### Added
+- Add blog_id to tracks data [#37902]
+
+## [2.10.2] - 2024-06-25
+### Changed
+- Internal updates.
+
 ## [2.10.1] - 2024-06-12
 ### Changed
 - Updated package dependencies. [#37796]
@@ -1115,6 +1123,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Separate the connection library into its own package.
 
+[2.11.0]: https://github.com/Automattic/jetpack-connection/compare/v2.10.2...v2.11.0
+[2.10.2]: https://github.com/Automattic/jetpack-connection/compare/v2.10.1...v2.10.2
 [2.10.1]: https://github.com/Automattic/jetpack-connection/compare/v2.10.0...v2.10.1
 [2.10.0]: https://github.com/Automattic/jetpack-connection/compare/v2.9.3...v2.10.0
 [2.9.3]: https://github.com/Automattic/jetpack-connection/compare/v2.9.2...v2.9.3