+++ projects/plugins/wpcomsh/CHANGELOG.md
@@ -1,5 +1,41 @@
 # Changelog
 
+## 3.26.1 - 2024-06-26
+### Added
+- Site Migrations: Add tracks events for AIOWPM events. [#37902]
+
+### Changed
+- Updated custom-fonts to v3.0.4. [#38071]
+
+## 3.26.0 - 2024-06-26
+### Added
+- Added polyfill for get_magic_quotes_gpc [#38039]
+
+## 3.25.2 - 2024-06-26
+### Changed
+- Internal updates.
+
+## 3.25.1 - 2024-06-25
+### Changed
+- Hide login banner for agency-managed sites. [#38021]
+
+### Removed
+- Plugin hotfixes: Remove a hotfix for an issue fixed in Gutenberg 15.3.0. [#38015]
+
+### Fixed
+- Custom Colors: Remove a use of `extract()`. [#38015]
+- Switch from Jetpack-the-plugin's `Jetpack_WPCOM_Block_Editor` class to jetpack-mu-wpcom's namespaced version. [#38015]
+
+## 3.25.0 - 2024-06-25
+### Added
+- Added a command to disable fatal error emails. [#38010]
+- Add `atomic_managed_plugin_row_auto_update_label` filter to translate the managed plugin auto update label [#37983]
+
+## 3.24.0 - 2024-06-24
+### Added
+- Added featrue flag checks for the Threads Social Connection. [#38001]
+- wpcomsh: Add WP-CLI commands with hooks for WoA post-transfer/reset/clone functionality [#37972]
+
 ## 3.23.0 - 2024-06-21
 ### Added
 - New wpcomsh CLI command to fix PHP 7.4 plugin [#37966]